Log package and process names on launch
When running on AndroidStudio, AndroidMonitor often reads the name of a new process before VirtualApp renames it.
Which means the process list will have virtualapp:p0, virtualapp:p1, etc -- Not very useful!

To make debugging easier, the packageName and processName are not one of the first things logged. I find it useful :)